LightCommand => CacheableCommand
