+++ autoload/clap/provider/grep.vim
@@ -252,7 +252,7 @@ if clap#maple#is_available()
   function! s:clear_job_and_matches() abort
   endfunction
 
-  function! s:start_job(query)
+  function! s:start_job(query) abort
     let [grep_opts, query] = s:translate_query_and_opts(a:query)
     " Add ' .' for windows in maple
     call clap#maple#command#start_grep_sync(s:grep_executable.' '.grep_opts, query, s:grep_enable_icon, s:ripgrep_glob)
